Update to include new revision
There are two versions of both firmwares (ESP32 and DCU). DCU version 1.2 or greater is designed to work with ESP32 version 3.1. All versions have an undated COmmander library that fixes an issue where sub commands that relied on the ON and OFF keywords did not work properly.

Both ESP32 versions support over the air firmware updates, however the process has proved unreliable. A method of updating the firmware using the DCU is being developed.